luaclass: Allow Lua to listen to global signal connections.
This is the successor of `awful.screen.connect_for_each_screen`-like
functions using the signal system itself "meta signals?". For signals
such as requests that need to be fullfilled at startup, the previous
API didn't "feel" native. Now it is possible for Lua to react to
global signal connections and call the handler itself.

This is intended to be used as a building block of a consistency
refactoring.

It could eventually be extended with an ugly global
`block_next_connection`. To allow filtering for signals that can
only have a single connection. So far it isn't required, so lets
not polute this patch with such hack.