travis: Pull the release from github.com using https://
www.lua.org is currently down and it may last for a while. They say on
IRC we can trust the github.com mirror as it is mostly official (but
maintained by the community and often off by a few days from the official
release)

The official mirrors are:

http://www.lua.org/ftp
http://webserver2.tecgraf.puc-rio.br/lua/mirror/ftp/

None have working https.