gears.delayed_call: Delay recursive delayed calls more
Image a function in a delayed call adds another delayed call.
Previously, this second delayed call was run immediately. This commit
changes things so that instead the second delayed call is moved to the
next main loop iteration.

It does that by collecting the list of delayed calls to run at the
beginning. Any newly incoming delayed calls thus cause another idle
callback to be added, which will only be serviced in a later main loop
iteration (=when no higher-priority event sources are pending).

Signed-off-by: Uli Schlachter <psychon@znc.in>