test-naughty-legacy: Use GDBus instead of spawning dbus-send
This gets rid of one more dependency, although awesome-client still
depends on dbus-send.

However, this is also a correctness fix: The "hints" argument to
notifications is supposed to have type a{sv}, i.e. this maps from string
keys to some generic value. dbus-send does not support this and instead
it sent an argument of type a{ss}, i.e. a mapping from string to string.
This commit fixes this problem.

Signed-off-by: Uli Schlachter <psychon@znc.in>