naughty.dbus: Fix inline icon data handling
LGI truncates GVariant bytestring instances at the first embedded \0
byte when using .value for "unwrapping". This commit works around that
problem by avoiding the .value API for accessing the image data [0].

Thanks a lot to Will Dietz for finding this problem and for providing a
preliminary patch fixing the problem. That saved me a lot of time [1].

[0]: https://github.com/pavouk/lgi/pull/223
[1]: https://github.com/dtzWill/awesome/commit/eecdeb7d469b061a2499eff3a60c6847d5cee020

Signed-off-by: Uli Schlachter <psychon@znc.in>