constrain awful.mouse.client.move/resize if maximized
I originally reported this bug as #2061, which got merged into #2036
which was reworded to apply to resize requests by the client *and* by
the user.  The issue however was closed without ever fixing the latter.
Hopefully this now finally fixes #2036 entirely.

This also reverts 51ddb5639e9c, which was another partial fix which is
unnecessary now and which had the undesirable side-effect of causing the
mouse to not be grabbed at all.