Drawin: Use a GLib idle source for resizes
Ever since the switch from libev to GLib, we have a hack in place. We
use g_main_context_set_poll_func() to replace g_poll() (which is just a
wrapper around poll()) with out own function that does lots of house
keeping. This commit is part of getting rid of this hack: It moves one
step that is done in this poll-replacement into a GLib idle source.

This commit should not be able to break any Lua code, but it might cause
visible difference, because the timing of when we resize drawins could
be slightly different than what we had before.

Signed-off-by: Uli Schlachter <psychon@znc.in>