Border updates: Use a GLib idle source
Ever since the switch from libev to GLib, we have a hack in place. We
use g_main_context_set_poll_func() to replace g_poll() (which is just a
wrapper around poll()) with our own function that does lots of house
keeping. This commit is part of getting rid of this hack: It moves one
step that is done in this poll-replacement into a GLib idle source.

This commit moves changing border widths to a GLib idle source. Instead
of doing it at the end of the main loop iteration, this is now properly
integrated with the main loop.

Note that this changes client_resize_do() to (implicitly) apply pending
border width updates. This is necessary since otherwise test-gravity
breaks. I guess that there always was a race between updating the border
width and the rest of the window geometry. If Awesome loses the race,
then a client can see its new geometry, but still have the old border
width. Just updating the border width at the same time as the rest of
the geometry makes this issue go away.

Signed-off-by: Uli Schlachter <psychon@znc.in>