Input focus: Use a GLib idle source for updates
Ever since the switch from libev to GLib, we have a hack in place. We
use g_main_context_set_poll_func() to replace g_poll() (which is just a
wrapper around poll()) with our own function that does lots of house
keeping. This commit is part of getting rid of this hack: It moves one
step that is done in this poll-replacement into a GLib idle source.

This commit moves updating the X11 input focus to an idle source.

The previous code unset globalconf.focus.need_update at the end of
client_focus_refresh() to work around client_unban() possibly setting it
to true again (see commit 56e727e026 for details). This commit splits up
client_unban() into two functions and only one of them will schedule a
focus update. This should fix the issue that is fixed by 56e727e026 as
well, so no regression should happen.

Signed-off-by: Uli Schlachter <psychon@znc.in>