Banning: Use a GLib idle source
Ever since the switch from libev to GLib, we have a hack in place. We
use g_main_context_set_poll_func() to replace g_poll() (which is just a
wrapper around poll()) with our own function that does lots of house
keeping. This commit is part of getting rid of this hack: It moves one
step that is done in this poll-replacement into a GLib idle source.

When a client becomes visible or invisible (e.g. because of a tag
switch), we have to tell the X11 server about it. Since the set of
visible clients can change a lot (e.g. first tag A is deselected, then
tag B is selected, then tag C is selected), we update this lazily (for
the example: Instead of doing three updates, only one is done).

This commit makes the straight-forward change necessary to move this to
an idle source.

Signed-off-by: Uli Schlachter <psychon@znc.in>