Stacking update: Use a GLib idle source
Ever since the switch from libev to GLib, we have a hack in place. We
use g_main_context_set_poll_func() to replace g_poll() (which is just a
wrapper around poll()) with our own function that does lots of house
keeping. This commit is part of getting rid of this hack: It moves one
step that is done in this poll-replacement into a GLib idle source.

Namely, this commit makes us use an idle source instead of a boolean to
track whether we have to update the stacking order.

Signed-off-by: Uli Schlachter <psychon@znc.in>