Use a GLib idle source for lazily destroying windows
Ever since the switch from libev to GLib, we have a hack in place. We
use g_main_context_set_poll_func() to replace g_poll() (which is just a
wrapper around poll()) with our own function that does lots of house
keeping. This commit is part of getting rid of this hack: It moves one
step that is done in this poll-replacement into a GLib idle source.

We lazily destroy frame windows, because as soon as the window is
destroyed, clients that now become visible get notified and repaint
themselves. However, it might be the case that we also need to update
the stacking order or unban clients, so that some other client is
visible in the end. To avoid this short flicker of "something else", we
lazily destroy some windows. This commit moves this into a GLib idle
source.

I did not add a variable that tracks whether we already added an idle
source for destroying windows. Instead, each time something is added to
the list, a new idle source is created. When it is dispatched, the first
source will handle the full list of windows to destroy. Later sources
will just do nothing. In my opinion, this is fine, because it is
unlikely that multiple windows end up on the to-be-destroyed list at the
same time.

Signed-off-by: Uli Schlachter <psychon@znc.in>