rc.lua: Switch from `awful.rules` to `ruled.client`.
It's the same module, but with a new name. The commit also switches
from the "one off" rule write to the incremental and signal based one.

This allows modules to add their own rules without having a race
condition against `rc.lua` and other modules. Previously, if a module
wasn't designed to prevent it, rules got destroyed each time.